Add files via upload
Exploratory Data Analysis is a very important stage in data analysis. It involves examining, and efforts to understand the dataset to know its characteristics, visualization is often employed in this process.
 Exploratory Data Analysis on the image dataset
slight shear transformations were applied, along with varying brightness between 90% and 110% of the original. These transformations help improve the robustness of the model. The rotation, shifts, zooms, flips, and shear transformations help to create multiple variations of the training images which increases the size of the training set without having to collect more data, which gives the model more images to train on. The randomness of the transformation makes the model robust which would help in real-world applications, the lighting transformations also help the model learn in different lighting conditions.Image visualization: To view a sample of the images in the benign and melanoma classes, a function ‘grid_display’ was defined to display the samples in a grid format to assess the quality and variety of images in the dataset.
Image color channels: Pixel intensity distributions across different color channels can be visualized to help identify patterns in different classes. The major color channels are Red, Green, Blue (primary colors). Picture intensity distribution can be used in quality assessment, feature extraction, and can also impact preprocessing distributions. The intensity values of the Red, green, and blue colors were plotted in this project to visualize the pixel intensity distribution of each class.